refactor(infra): move QueueNames from core to module-infra
Issue #906: core port interfaces no longer expose PGMQ queue name strings.
Adapter-side constants live in module-infra/infrastructure/queue/. All
10 callers updated to new import path. Compile clean across all modules.
