+++ module-app/src/main/java/maple/expectation/application/service/expectation/queue/ExpectationCalculationQueue.java
@@ -7,7 +7,7 @@ import maple.expectation.core.model.job.CalculationJobStatus;
 import maple.expectation.core.port.inbound.TaskReceipt;
 import maple.expectation.core.port.out.CalculationJobPort;
 import maple.expectation.core.port.out.MessageQueuePort;
-import maple.expectation.core.port.out.QueueNames;
+import maple.expectation.infrastructure.queue.QueueNames;
 import maple.expectation.infrastructure.executor.LogicExecutor;
 import maple.expectation.infrastructure.executor.TaskContext;
 import maple.expectation.infrastructure.pgmq.ExternalApiJobPayload;
